requirements.txt
faker==37.1.0
fastapi==0.115.12
numpy==2.2.4
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.1.0
scikit-learn==1.6.1
sqlalchemy==2.0.40
uvicorn==0.34.0
